requirements.txt
affine==2.3.0
aniso8601==8.0.0
attrs==19.3.0
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==6.7
click-plugins==1.1.1
cligj==0.5.0
cryptography==2.9.2
Flask==0.12.2
Flask-API==1.0
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
idna==2.9
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
numpy==1.18.2
opencv-python==4.2.0.32
paramiko==2.7.1
Pillow==7.0.0
progressbar==2.5
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.6
pyproj==2.6.1.post1
pysftp==0.2.9
pytz==2019.3
rasterio==1.1.3
requests==2.23.0
six==1.14.0
snuggs==1.4.7
torch==1.4.0
torchvision==0.5.0
urllib3==1.25.8
Werkzeug==0.12.2
python-dateutil~=2.8.1